+++ src/udm_app/udm_app.hpp
@@ -60,14 +60,14 @@ class udm_app {
    * @param [const oai::model::udm::AuthenticationInfoRequest&]
    * authenticationInfoRequest: request's info
    * @param [nlohmann::json&] auth_info_response: Authentication response's info
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @return void
    */
   void handle_generate_auth_data_request(
       const std::string& supiOrSuci,
       const oai::model::udm::AuthenticationInfoRequest&
           authenticationInfoRequest,
-      nlohmann::json& auth_info_response, long& code);
+      nlohmann::json& auth_info_response, uint32_t& code);
 
   /*
    * Handle a request to confirm the authentication data
@@ -75,12 +75,12 @@ class udm_app {
    * @param [const oai::model::udm::AuthEvent&] authEvent: Authentication Event
    * @param [nlohmann::json&] confirm_response: Confirm response
    * @param [std::string&] location: location of the resource
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @return void
    */
   void handle_confirm_auth(
       const std::string& supi, const oai::model::udm::AuthEvent& authEvent,
-      nlohmann::json& confirm_response, std::string& location, long& code);
+      nlohmann::json& confirm_response, std::string& location, uint32_t& code);
 
   /*
    * Handle a request to delete an authentication data
@@ -88,24 +88,24 @@ class udm_app {
    * @param [const std::string&] authEventId: Event ID
    * @param [const oai::model::udm::AuthEvent&] authEvent: Authentication Event
    * @param [nlohmann::json&] auth_response: Authentication response
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @return void
    */
   void handle_delete_auth(
       const std::string& supi, const std::string& authEventId,
       const oai::model::udm::AuthEvent& authEvent,
-      nlohmann::json& auth_response, long& code);
+      nlohmann::json& auth_response, uint32_t& code);
 
   /*
    * Handle a request to get the Access and Mobility Subscription Data
    * @param [const std::string&] supi: UE's SUPI
    * @param [nlohmann::json&] response_data: response's info
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @param [oai::model::common::PlmnId] PlmnId: PLMN ID
    * @return void
    */
   void handle_access_mobility_subscription_data_retrieval(
-      const std::string& supi, nlohmann::json& response_data, long& code,
+      const std::string& supi, nlohmann::json& response_data, uint32_t& code,
       oai::model::common::PlmnId PlmnId = {});
 
   /*
@@ -114,7 +114,7 @@ class udm_app {
    * @param [const oai::model::udm::Amf3GppAccessRegistration&]
    * amf_3gpp_access_registration: Registration info
    * @param [nlohmann::json&] response_data: response's info
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @param [oai::model::common::PlmnId] PlmnId: PLMN ID
    * @return void
    */
@@ -122,20 +122,20 @@ class udm_app {
       const std::string& ue_id,
       const oai::model::udm::Amf3GppAccessRegistration&
           amf_3gpp_access_registration,
-      nlohmann::json& response_data, long& code);
+      nlohmann::json& response_data, uint32_t& code);
 
   /*
    * Handle a request to get the Session Management Subscription Data
    * @param [const std::string&] supi: UE's SUPI
    * @param [nlohmann::json&] response_data: response's info
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @param [oai::model::common::Snssai] snssai: SNSSAI
    * @param [oai::model::common::PlmnId] PlmnId: PLMN ID
    * @param [std::string&] dnn: DNN
    * @return void
    */
   void handle_session_management_subscription_data_retrieval(
-      const std::string& supi, nlohmann::json& response_data, long& code,
+      const std::string& supi, nlohmann::json& response_data, uint32_t& code,
       oai::model::common::Snssai snssai = {}, std::string dnn = {},
       oai::model::common::PlmnId plmn_id = {});
 
@@ -143,12 +143,12 @@ class udm_app {
    * Handle a request to get the Slice Selection Subscription Data
    * @param [const std::string&] supi: UE's SUPI
    * @param [nlohmann::json&] response_data: response's info
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @param [oai::model::common::PlmnId] PlmnId: PLMN ID
    * @return void
    */
   void handle_slice_selection_subscription_data_retrieval(
-      const std::string& supi, nlohmann::json& response_data, long& code,
+      const std::string& supi, nlohmann::json& response_data, uint32_t& code,
       std::string supported_features     = {},
       oai::model::common::PlmnId plmn_id = {});
 
@@ -156,13 +156,13 @@ class udm_app {
    * Handle a request to get the SMF Selection Subscription Data
    * @param [const std::string&] supi: UE's SUPI
    * @param [nlohmann::json&] response_data: response's info
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @param [std::string] supported_features: supported features
    * @param [oai::model::common::PlmnId] PlmnId: PLMN ID
    * @return void
    */
   void handle_smf_selection_subscription_data_retrieval(
-      const std::string& supi, nlohmann::json& response_data, long& code,
+      const std::string& supi, nlohmann::json& response_data, uint32_t& code,
       std::string supported_features     = {},
       oai::model::common::PlmnId plmn_id = {});
 
@@ -172,13 +172,13 @@ class udm_app {
    * @param [const oai::model::udm::SdmSubscription&] sdmSubscription:
    * Suscription info
    * @param [nlohmann::json&] response_data: response's info
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @return void
    */
   void handle_subscription_creation(
       const std::string& supi,
       const oai::model::udm::SdmSubscription& sdmSubscription,
-      nlohmann::json& response_data, long& code);
+      nlohmann::json& response_data, uint32_t& code);
 
   /*
    * Handle a request to create an event subscription
@@ -187,13 +187,13 @@ class udm_app {
    * info
    * @param [const oai::model::udm::CreatedEeSubscription&] createdSub: created
    * suscription info
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @return subscription Id
    */
   evsub_id_t handle_create_ee_subscription(
       const std::string& ueIdentity,
       const oai::model::udm::EeSubscription& eeSubscription,
-      oai::model::udm::CreatedEeSubscription& createdSub, long& code);
+      oai::model::udm::CreatedEeSubscription& createdSub, uint32_t& code);
 
   /*
    * Handle a request to delete an event subscription
@@ -201,12 +201,12 @@ class udm_app {
    * @param [const std::string&] subscriptionId: subscription's Id
    * @param [oai::model::common::ProblemDetails&] problemDetails: problem
    * happened (if exist) when deleting the even
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @return void
    */
   void handle_delete_ee_subscription(
       const std::string& ueIdentity, const std::string& subscriptionId,
-      oai::model::common::ProblemDetails& problemDetails, long& code);
+      oai::model::common::ProblemDetails& problemDetails, uint32_t& code);
 
   /*
    * Handle a request to update an event subscription
@@ -216,13 +216,13 @@ class udm_app {
    * of actions
    * @param [oai::model::common::ProblemDetails&] problemDetails: problem
    * happened (if exist) when executing the requests
-   * @param [long&] code: response's code
+   * @param [uint32_t&] code: response's code
    * @return void
    */
   void handle_update_ee_subscription(
       const std::string& ueIdentity, const std::string& subscriptionId,
       const std::vector<oai::model::common::PatchItem>& patchItem,
-      oai::model::common::ProblemDetails& problemDetails, long& code);
+      oai::model::common::ProblemDetails& problemDetails, uint32_t& code);
 
   /*
    * Generate an unique ID for the new subscription
